put clean architecture + change warp to use rocket framework

+++ src/domain/company/model/company.rs
@@ -1,12 +1,14 @@
+use chrono::NaiveDateTime;
+
 pub struct Company {
-    pub id: Option<i32>,
+    pub id: Option<u64>,
     pub name: String,
     pub description: String,
     pub address: String,
-    pub post_code: i8,
+    pub post_code: i32,
     pub city: String,
     pub country: String,
     pub abn: Option<String>,
-    pub ceo_id: Option<i32>,
+    pub ceo_id: Option<u64>,
     pub created_at: NaiveDateTime,
 }
